use own server-destroy implementation that supports secureConnect events
