+++ packages/network/lib/index.ts
@@ -1,6 +1,9 @@
 import agent from './agent'
 import * as connect from './connect'
+import { allowDestroy } from './allow-destroy'
 
-export { agent }
-
-export { connect }
+export {
+  agent,
+  allowDestroy,
+  connect,
+}
