+++ packages/server/lib/util/server_destroy.js
@@ -1,12 +1,5 @@
-// TODO: This file was created by bulk-decaffeinate.
-// Sanity-check the conversion and remove this comment.
-/*
- * decaffeinate suggestions:
- * DS102: Remove unnecessary code created because of implicit returns
- * Full docs: https://github.com/decaffeinate/decaffeinate/blob/master/docs/suggestions.md
- */
 const Promise = require('bluebird')
-const allowDestroy = require('server-destroy')
+const { allowDestroy } = require('@packages/network')
 
 module.exports = function (server) {
   allowDestroy(server)
@@ -16,4 +9,4 @@ module.exports = function (server) {
     .catch(() => {})
   }
 }
-//# dont catch any errors
+// dont catch any errors
