+++ packages/server/package.json
@@ -118,7 +118,6 @@
     "sanitize-filename": "1.6.1",
     "semver": "6.3.0",
     "send": "0.17.0",
-    "server-destroy": "1.0.1",
     "shell-env": "3.0.0",
     "signal-exit": "3.0.2",
     "sinon": "5.1.1",
